# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Foohut Documentation Platform API
  version: 1.0.0
  description: |
    AI-native documentation platform API with hierarchical content organization,
    block-based editing, Git synchronization, and AI-powered features.

    ## Architecture
    - **Organization** > **Collection** > **Space** > **Page** > **Block**
    - JWT-based authentication with role-based access control
    - Cursor-based pagination for list endpoints
    - Rate limiting with standard headers

    ## Authentication
    All endpoints (except auth) require a valid JWT Bearer token.
    Include the token in the Authorization header: `Bearer <token>`
  contact:
    name: Foohut API Support
    email: api@foohut.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.foohut.com/v1
    description: Production server
  - url: https://staging-api.foohut.com/v1
    description: Staging server
  - url: http://localhost:3000/v1
    description: Local development
...
security:
  - BearerAuth: []
paths:
...
  /organizations:
    get:
      tags:
        - Organizations
      summary: List organizations
      description: List all organizations the authenticated user has access to
      operationId: listOrganizations
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Organizations retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Organizations
      summary: Create organization
      description: Create a new organization
      operationId: createOrganization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrganizationRequest'
            example:
              name: Acme Corp
              slug: acme-corp
              description: Documentation for Acme Corporation
      responses:
        '201':
          description: Organization created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Organization slug already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /organizations/{orgId}:
    get:
      tags:
        - Organizations
      summary: Get organization
      description: Retrieve organization details by ID
      operationId: getOrganization
      parameters:
        - $ref: '#/components/parameters/OrgId'
      responses:
        '200':
          description: Organization retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags:
        - Organizations
      summary: Update organization
      description: Update organization details
      operationId: updateOrganization
      parameters:
        - $ref: '#/components/parameters/OrgId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrganizationRequest'
      responses:
        '200':
          description: Organization updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Organizations
      summary: Delete organization
      description: Permanently delete an organization and all its content
      operationId: deleteOrganization
      parameters:
        - $ref: '#/components/parameters/OrgId'
      responses:
        '204':
          description: Organization deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /organizations/{orgId}/members:
    get:
      tags:
        - Organizations
      summary: List organization members
      description: List all members of an organization
      operationId: listOrganizationMembers
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: role
          in: query
          description: Filter by role
          schema:
            $ref: '#/components/schemas/OrganizationRole'
      responses:
        '200':
          description: Members retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags:
        - Organizations
      summary: Add organization member
      description: Invite a user to join the organization
      operationId: addOrganizationMember
      parameters:
        - $ref: '#/components/parameters/OrgId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddMemberRequest'
            example:
              email: newmember@example.com
              role: editor
      responses:
        '201':
          description: Member added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: User is already a member
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /organizations/{orgId}/members/{userId}:
    patch:
      tags:
        - Organizations
      summary: Update member role
      description: Update a member's role in the organization
      operationId: updateOrganizationMember
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - role
              properties:
                role:
                  $ref: '#/components/schemas/OrganizationRole'
      responses:
        '200':
          description: Member role updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Organizations
      summary: Remove organization member
      description: Remove a member from the organization
      operationId: removeOrganizationMember
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - $ref: '#/components/parameters/UserId'
      responses:
        '204':
          description: Member removed successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /organizations/{orgId}/collections:
    get:
      tags:
        - Collections
      summary: List collections
      description: List all collections in an organization
      operationId: listCollections
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Collections retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags:
        - Collections
      summary: Create collection
      description: Create a new collection in an organization
      operationId: createCollection
      parameters:
        - $ref: '#/components/parameters/OrgId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCollectionRequest'
            example:
              name: Product Documentation
              slug: product-docs
              description: Technical documentation for our product suite
              visibility: public
      responses:
        '201':
          description: Collection created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Collection'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Collection slug already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /auth/login
  parameters:
    UserId:
      name: userId
      in: path
      required: true
      description: User unique identifier
      schema:
        type: string
        format: uuid
    OrgId:
      name: orgId
      in: path
      required: true
      description: Organization unique identifier
      schema:
        type: string
        format: uuid
...
    Cursor:
      name: cursor
      in: query
      description: Pagination cursor for next page
      schema:
        type: string
    Limit:
      name: limit
      in: query
      description: Maximum number of items to return
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: BAD_REQUEST
              message: Validation failed
              details:
                - field: email
                  message: Invalid email format
    Unauthorized:
      description: Unauthorized - invalid or missing authentication
      headers:
        WWW-Authenticate:
          schema:
            type: string
          example: Bearer realm="api"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: UNAUTHORIZED
              message: Invalid or expired token
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: FORBIDDEN
              message: You do not have permission to perform this action
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: NOT_FOUND
              message: Resource not found
...
  schemas:
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Error code for programmatic handling
              enum:
                - BAD_REQUEST
                - UNAUTHORIZED
                - FORBIDDEN
                - NOT_FOUND
                - CONFLICT
                - RATE_LIMITED
                - INTERNAL_ERROR
            message:
              type: string
              description: Human-readable error message
            details:
              type: array
              description: Additional error details
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
            requestId:
              type: string
              description: Request ID for support reference
...
    PaginationMeta:
      type: object
      properties:
        cursor:
          type: string
          nullable: true
          description: Cursor for the next page (null if no more pages)
        hasMore:
          type: boolean
          description: Whether more items exist
        total:
          type: integer
          description: Total count of items (when available)

    # --------------------------------------------------------------------------
    # AUTH SCHEMAS
    # --------------------------------------------------------------------------
...
    UserPublic:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        avatarUrl:
          type: string
          format: uri
          nullable: true
...
    Organization:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
          pattern: ^[a-z0-9-]+$
        description:
          type: string
          nullable: true
        logoUrl:
          type: string
          format: uri
          nullable: true
        ownerId:
          type: string
          format: uuid
        plan:
          type: string
          enum: [free, pro, enterprise]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    OrganizationList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Organization'
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
    CreateOrganizationRequest:
      type: object
      required:
        - name
        - slug
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        slug:
          type: string
          pattern: ^[a-z0-9-]+$
          minLength: 3
          maxLength: 50
        description:
          type: string
          maxLength: 500
    UpdateOrganizationRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 500
        logoUrl:
          type: string
          format: uri
          nullable: true
    OrganizationRole:
      type: string
      enum: [owner, admin, editor, viewer]
    Member:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        user:
          $ref: '#/components/schemas/UserPublic'
        role:
          $ref: '#/components/schemas/OrganizationRole'
        joinedAt:
          type: string
          format: date-time
    MemberList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Member'
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
    AddMemberRequest:
      type: object
      required:
        - email
        - role
      properties:
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/OrganizationRole'
...
    Collection:
      type: object
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
          pattern: ^[a-z0-9-]+$
        description:
          type: string
          nullable: true
        visibility:
          type: string
          enum: [public, private]
        coverImageUrl:
          type: string
          format: uri
          nullable: true
        spacesCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CollectionList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Collection'
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
    CreateCollectionRequest:
      type: object
      required:
        - name
        - slug
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        slug:
          type: string
          pattern: ^[a-z0-9-]+$
          minLength: 3
          maxLength: 50
        description:
          type: string
          maxLength: 500
        visibility:
          type: string
          enum: [public, private]
          default: private
...